+++ src-tauri/src/mods/proveedor.rs
@@ -8,13 +8,13 @@ use super::lib::Save;
 
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 pub struct Proveedor {
-    id: i32,
+    id: i64,
     nombre: Arc<str>,
     contacto: Option<i64>,
 }
 
 impl Proveedor {
-    pub fn new(id: i32, nombre: &str, contacto: Option<i64>) -> Self {
+    pub fn new(id: i64, nombre: &str, contacto: Option<i64>) -> Self {
         Proveedor {
             id,
             nombre: Arc::from(nombre),
@@ -24,7 +24,7 @@ impl Proveedor {
     pub fn nombre(&self) -> Arc<str> {
         Arc::clone(&self.nombre)
     }
-    pub fn id(&self) -> &i32 {
+    pub fn id(&self) -> &i64 {
         &self.id
     }
     pub fn contacto(&self) -> &Option<i64> {
