+++ src-tauri/src/mods/relacion_prod_prov.rs
@@ -6,23 +6,23 @@ use super::lib::Save;
 
 #[derive(Clone, Serialize, Deserialize)]
 pub struct RelacionProdProv {
-    id_producto: i32,
-    id_proveedor: i32,
+    id_producto: i64,
+    id_proveedor: i64,
     codigo_interno: Option<i64>,
 }
 
 impl RelacionProdProv {
-    pub fn new(id_producto: i32, id_proveedor: i32, codigo_interno: Option<i64>) -> Self {
+    pub fn new(id_producto: i64, id_proveedor: i64, codigo_interno: Option<i64>) -> Self {
         RelacionProdProv {
             id_producto,
             id_proveedor,
             codigo_interno,
         }
     }
-    pub fn id_producto(&self) -> &i32 {
+    pub fn id_producto(&self) -> &i64 {
         &self.id_producto
     }
-    pub fn id_proveedor(&self) -> &i32 {
+    pub fn id_proveedor(&self) -> &i64 {
         &self.id_proveedor
     }
     pub fn codigo_interno(&self) -> Option<i64> {
@@ -32,8 +32,8 @@ impl RelacionProdProv {
 impl Save for RelacionProdProv {
     async fn save(&self) -> Result<(), DbErr> {
         let model = relacion_prod_prov::ActiveModel {
-            producto: Set(*self.id_producto() as i64),
-            proveedor: Set(*self.id_producto() as i64),
+            producto: Set(*self.id_producto()),
+            proveedor: Set(*self.id_producto()),
             codigo: Set(self.codigo_interno),
             ..Default::default()
         };
